final stages of repo cleanup
