+++ src/plotBot/command.js
@@ -184,3 +184,36 @@ function CreateCommand(messageContent) {
         return null;
     }
 }
+
+function getHelp() {
+    var helpText;
+    helpText += "Hey there! My name is PlotBot! :sunglasses:\n";
+    helpText += "I'm a discord bot that helps users plot and track values and trends over time.\n";
+    helpText += "First, plot a few points using the `add` command. Like this:\n";
+    helpText += "`@PlotBot add 75`\n";
+    helpText += "This will create a plot point for your user with value of 75 for this instance in time.\n";
+    helpText += "You can also specify other users by @-ing them after you specify the value to plot for them:\n";
+    helpText += "`@PlotBot add 124 @userName`\n";
+    helpText += "This will create a plot point for their user with value of 124 for this instance in time.\n";
+
+    // no ability to specify custom moments for now.
+    /*
+    helpText += "Finally, you can specify a point in time other than right now by specifying a date and time string after the user:\n";
+    helpText += "`@PlotBot add 394 @userName May 4th 2020 3:30pm`\n";
+    helpText += "This will create a plot point for that user with value of 394 on May 4th at 3:30.\n";
+    helpText += "If you want to specify a plot for your own graph but at a different time, just @ your own user.\n";
+    */
+
+    helpText += "Now that you have given me some data, I can do some cool stuff with it: \n";
+    helpText += "`@PlotBot graph week`\n";
+    helpText += "Then I'll send you a graph of all plots over the last week for your user that looks like this:\n";
+    helpText += "<placeholder sample graph image>\n";
+
+    helpText += "You can specify day, week, month or year. You can also add additional users you would like to see graphed along with yourself.\n";
+    helpText += "`@PlotBot graph month @userName @wumpus`\n";
+    helpText += "<placeholder sample graph image>\n";
+
+    helpText += "Thanks for listening! I can't wait to get to work! :smile: ";
+
+    return helpText;
+}
